pipeline/airflow/requirements.txt
numpy
pandas
twint
nltk
psycopg2-binary
sqlalchemy<2.0
apache-airflow-providers-postgres
tensorflow